requirements.txt
setuptools==59.5.*
opencv-python
opencv-contrib-python
prettytable
tqdm
yacs
torchfile
plyfile
munkres
chumpy
trimesh
shyaml
PyYAML==5.4.1
numpy==1.22.0
numpy-quaternion
pygame
keyboard
transforms3d
bvhtoolbox
vedo
imgaug==0.4.0
tensorboard
smplx
pycocotools
plotly
open3d
addict
loguru
norfair
h5py
imageio-ffmpeg
pyrender
joblib
pandas